fix update handler

+++ src/guide/service.py
@@ -1,5 +1,5 @@
 from typing import List
-from fastapi import Depends
+from fastapi import Depends, HTTPException, status
 from sqlmodel import Session, select
 from config.db import get_session
 from .model import Guide, GuideModel
@@ -25,30 +25,31 @@ async def create(guide: GuideModel, session: Session = Depends(get_session)) -> 
 
 
 async def update(
-    guide_id: str, guide: GuideModel, session: Session = Depends(get_session)
+    guide_id: str, author_id: str, guide: GuideModel, session: Session = Depends(get_session)
 ) -> Guide:
     try:
-        statement = session.exec(
-            select(GuideModel).where(GuideModel.id == int(guide_id))
-        ).first()
-        GuideModel.model_validate(guide)
-        statement.title = guide["title"]
-        statement.description = guide["description"]
-        statement.img = guide["img"]
-        statement.author = guide["author"]
-        session.add(statement)
+        query = select(GuideModel).where(GuideModel.id == int(guide_id))
+        obj = session.exec(query).first()
+        if not obj:
+            raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Note not found")
+        if obj.author != author_id:
+            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="You are unauthorized to perform this task.")
+        obj.sqlmodel_update(guide)
+        session.add(obj)
         session.commit()
-        session.refresh(statement)
-        return statement
+        session.refresh(obj)
+        return obj
     except Exception as e:
         raise e
 
 
-async def delete_gui(guide_id: str, session: Session = Depends(get_session)) -> None:
+async def delete_gui(guide_id: str, author_id: str, session: Session = Depends(get_session)) -> None:
     try:
         statement = session.exec(
             select(GuideModel).where(GuideModel.id == int(guide_id))
         ).first()
+        if statement.author != author_id:
+            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="You are unauthorized to perform this task.")
         session.delete(statement)
         session.commit()
     except Exception as e:
